Buttons EXTI callbacks, date and time displaying
+ implemented buttons EXTI Callback
+ implemented functions displaying date and time

+++ HAL_VFD_Clock/Core/Inc/display.h
@@ -52,6 +52,6 @@ void multiplexerSequence();
 uint8_t charToBitmask(char inputChar);
 void dispString(const char *string, uint8_t pos);
 void displayHour(uint8_t hour, uint8_t minute, uint8_t second);
-void displayDate(uint8_t day, uint8_t month, uint16_t year);
+void displayDate(uint8_t day, uint8_t month, uint8_t year);
 
 #endif /* INC_DISPLAY_H_ */
